# openapi.yaml
openapi: 3.0.3
info:
  title: invenchecker API
  version: 0.0.1
  description: Steam inventory checker and price alert service

servers:
  - url: http://localhost:33001

tags:
  - name: health
  - name: accounts
  - name: alerts

paths:
  /health:
    get:
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  lastScannedAt:
                    type: integer
                    nullable: true
                    description: Unix timestamp of last scan
                  lastScanMs:
                    type: integer
                    nullable: true
                    description: Duration of last scan in milliseconds
                  inventoryQueueSize:
                    type: integer
                    description: Number of steam64ids pending in the inventory queue
                  priceQueueSize:
                    type: integer
                    description: Number of items pending in the price queue

  /accounts:
    get:
      tags: [accounts]
      summary: List all accounts
      responses:
        "200":
          description: Array of all accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"

    post:
      tags: [accounts]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [friendlyName, discordId, steam64ids]
              properties:
                friendlyName:
                  type: string
                discordId:
                  type: string
                steam64ids:
                  type: array
                  items:
                    type: string
                    pattern: '^7656119\d{10}$'
                    description: 17-digit Steam64 ID starting with 7656119
                  minItems: 1
                customItems:
                  type: array
                  items:
                    type: string
                  default: []
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Missing or invalid fields, or entry previously rejected by the scanner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          $ref: "#/components/responses/Conflict"

  /accounts/discord:
    post:
      tags: [accounts]
      summary: Create account via Discord ID (minimal)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [discordId]
              properties:
                discordId:
                  type: string
                friendlyName:
                  type: string
                  nullable: true
      responses:
        "201":
          description: Account created, returns uid
          content:
            application/json:
              schema:
                type: object
                properties:
                  uid:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"

  /accounts/{uid}:
    parameters:
      - $ref: "#/components/parameters/uid"

    get:
      tags: [accounts]
      summary: Get account by UID
      responses:
        "200":
          description: Account found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [accounts]
      summary: Update account fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                friendlyName:
                  type: string
                discordId:
                  type: string
                steam64ids:
                  type: array
                  items:
                    type: string
                    pattern: '^7656119\d{10}$'
                    description: 17-digit Steam64 ID starting with 7656119
                customItems:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Invalid fields or entry previously rejected by the scanner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [accounts]
      summary: Delete account
      responses:
        "204":
          description: Account deleted
        "404":
          $ref: "#/components/responses/NotFound"

  /accounts/{uid}/steam64ids:
    parameters:
      - $ref: "#/components/parameters/uid"

    post:
      tags: [accounts]
      summary: Add a Steam64 ID to account (no-op if already present)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [steam64id]
              properties:
                steam64id:
                  type: string
                  pattern: '^7656119\d{10}$'
                  description: 17-digit Steam64 ID starting with 7656119
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Invalid or previously rejected steam64id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"

  /accounts/{uid}/steam64ids/{id}:
    parameters:
      - $ref: "#/components/parameters/uid"
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Steam64 ID to remove

    delete:
      tags: [accounts]
      summary: Remove a Steam64 ID from account
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          $ref: "#/components/responses/NotFound"

  /accounts/{uid}/customItems:
    parameters:
      - $ref: "#/components/parameters/uid"

    post:
      tags: [accounts]
      summary: Add a custom item to account (no-op if already present)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [item]
              properties:
                item:
                  type: string
                  description: market_hash_name of the item
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Item previously rejected by the scanner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"

  /accounts/{uid}/customItems/{item}:
    parameters:
      - $ref: "#/components/parameters/uid"
      - name: item
        in: path
        required: true
        schema:
          type: string
        description: URL-encoded market_hash_name to remove

    delete:
      tags: [accounts]
      summary: Remove a custom item from account
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          $ref: "#/components/responses/NotFound"

  /accounts/{uid}/inventory:
    parameters:
      - $ref: "#/components/parameters/uid"

    get:
      tags: [accounts]
      summary: Live Steam inventory for all steam64ids on the account
      responses:
        "200":
          description: Merged inventory items from all Steam accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  uid:
                    type: string
                  count:
                    type: integer
                  items:
                    type: array
                    items:
                      type: object
                      description: Raw Steam inventory item
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Steam API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /accounts/{uid}/summary:
    parameters:
      - $ref: "#/components/parameters/uid"

    get:
      tags: [accounts]
      summary: Inventory summary with latest prices per item
      responses:
        "200":
          description: Summary of tracked items and custom items with latest prices
          content:
            application/json:
              schema:
                type: object
                properties:
                  uid:
                    type: string
                  friendlyName:
                    type: string
                    nullable: true
                  steam64ids:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        $ref: "#/components/schemas/InventoryItem"
                  customItems:
                    type: array
                    items:
                      $ref: "#/components/schemas/CustomItem"
        "404":
          $ref: "#/components/responses/NotFound"

  /accounts/{uid}/progress:
    parameters:
      - $ref: "#/components/parameters/uid"

    get:
      tags: [accounts]
      summary: Scan state per steam64id and custom item
      description: Shows whether each steam64id/item is queued, when it was last fetched/priced, and when it is next scheduled to scan.
      responses:
        "200":
          description: Progress for each steam64id and custom item
          content:
            application/json:
              schema:
                type: object
                properties:
                  uid:
                    type: string
                  steam64ids:
                    type: object
                    additionalProperties:
                      $ref: "#/components/schemas/Steam64Progress"
                  customItems:
                    type: object
                    additionalProperties:
                      $ref: "#/components/schemas/ItemProgress"
        "404":
          $ref: "#/components/responses/NotFound"

  /accounts/{uid}/prices:
    parameters:
      - $ref: "#/components/parameters/uid"
      - name: days
        in: query
        schema:
          type: integer
          default: 7
        description: Number of days of price history to return
      - name: item
        in: query
        schema:
          type: string
        description: Filter to a single item by market_hash_name; defaults to account's customItems

    get:
      tags: [accounts]
      summary: Price history for account's custom items
      responses:
        "200":
          description: Price snapshots grouped by market_hash_name
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: "#/components/schemas/PriceSnapshot"
        "404":
          $ref: "#/components/responses/NotFound"

  /alerts:
    get:
      tags: [alerts]
      summary: List all alerts (admin)
      responses:
        "200":
          description: All alerts ordered by created_at desc
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"

  /alerts/user/{uid}:
    parameters:
      - $ref: "#/components/parameters/uid"

    get:
      tags: [alerts]
      summary: Get unresolved alerts for a user
      responses:
        "200":
          description: Unresolved alert recipients for the given uid
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertRecipient"

  /alerts/recipients/{id}/resolve:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: Alert recipient row ID

    put:
      tags: [alerts]
      summary: Resolve a single alert recipient
      responses:
        "200":
          description: Resolved recipient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRecipient"
        "404":
          $ref: "#/components/responses/NotFound"

  /alerts/user/{uid}/resolve-all:
    parameters:
      - $ref: "#/components/parameters/uid"

    put:
      tags: [alerts]
      summary: Resolve all unresolved alerts for a user
      responses:
        "200":
          description: Number of rows resolved
          content:
            application/json:
              schema:
                type: object
                properties:
                  resolved:
                    type: integer

  /alerts/scan:
    post:
      tags: [alerts]
      summary: Enqueue all accounts for scanning immediately (returns instantly)
      parameters:
        - name: force
          in: query
          schema:
            type: boolean
            default: false
          description: Bypass recency checks and enqueue all items immediately regardless of last scan time
      responses:
        "200":
          description: All accounts enqueued for scanning
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Scan enqueued
                  force:
                    type: boolean
                    description: Whether recency checks were bypassed
                  queuedAt:
                    type: integer
                    description: Unix timestamp when scan was enqueued

components:
  parameters:
    uid:
      name: uid
      in: path
      required: true
      schema:
        type: string
      description: Account UID (hex string)

  responses:
    BadRequest:
      description: Missing or invalid request fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string

    Account:
      type: object
      properties:
        uid:
          type: string
        friendlyName:
          type: string
          nullable: true
        discordId:
          type: string
        steam64ids:
          type: array
          items:
            type: string
            pattern: '^7656119\d{10}$'
        customItems:
          type: array
          items:
            type: string

    PriceSnapshot:
      type: object
      properties:
        lowest_price:
          type: number
          nullable: true
        median_price:
          type: number
          nullable: true
        volume:
          type: integer
          nullable: true
        captured_at:
          type: integer
          description: Unix timestamp

    InventoryItem:
      type: object
      properties:
        market_hash_name:
          type: string
        first_seen:
          type: integer
          description: Unix timestamp
        last_seen:
          type: integer
          description: Unix timestamp
        missing:
          type: boolean
        price:
          $ref: "#/components/schemas/PriceSnapshot"
          nullable: true

    CustomItem:
      type: object
      properties:
        market_hash_name:
          type: string
        price:
          $ref: "#/components/schemas/PriceSnapshot"
          nullable: true

    Alert:
      type: object
      properties:
        id:
          type: integer
        market_hash_name:
          type: string
        spike_pct:
          type: number
        price_at_alert:
          type: number
        seven_day_low:
          type: number
        created_at:
          type: integer
          description: Unix timestamp

    AlertRecipient:
      allOf:
        - $ref: "#/components/schemas/Alert"
        - type: object
          properties:
            recipient_id:
              type: integer
            resolved:
              type: integer
              enum: [0, 1]
            resolved_at:
              type: integer
              nullable: true
              description: Unix timestamp

    InventoryFetch:
      type: object
      nullable: true
      properties:
        item_count:
          type: integer
          description: Number of items found in the inventory
        duration_ms:
          type: integer
          description: Time taken to fetch the inventory in milliseconds
        fetched_at:
          type: integer
          description: Unix timestamp of the fetch

    Steam64Progress:
      type: object
      properties:
        queued:
          type: boolean
          description: True if this steam64id is currently pending in the inventory queue
        lastFetch:
          $ref: "#/components/schemas/InventoryFetch"
        nextScanAt:
          type: integer
          nullable: true
          description: Estimated Unix timestamp of next scan (lastFetch.fetched_at + re-enqueue delay), or null if queued or never scanned

    ItemProgress:
      type: object
      properties:
        queued:
          type: boolean
          description: True if this item is currently pending in the price queue
        lastPrice:
          type: object
          nullable: true
          properties:
            lowest_price:
              type: number
            captured_at:
              type: integer
              description: Unix timestamp
        nextScanAt:
          type: integer
          nullable: true
          description: Estimated Unix timestamp of next price check, or null if queued or never scanned