# src/routes/accounts.js
"use strict";

const crypto = require("crypto");
const express = require("express");
const router = express.Router();
const { readConfig, writeConfig } = require("../config");
const { fetchInventory } = require("../steam");
const db = require("../db");
const logger = require("../logger");
const { enqueueInventoryIfDue, enqueuePrice, isInventoryQueued, isPriceQueued, getQueueState } = require("../queue");
const { MAX_STEAM64IDS, MAX_CUSTOM_ITEMS, REENQUEUE_DELAY_MS } = require("../appConfig");

function getAccount(uid) {
  const accounts = readConfig();
  const account = accounts.find((a) => a.uid === uid);
  return { accounts, account };
}

function isValidSteam64id(id) {
  return typeof id === 'string' && /^7656119\d{10}$/.test(id);
}

function getBadEntryReason(type, value) {
  return db.prepare('SELECT reason FROM bad_entries WHERE type = ? AND value = ?').get(type, value)?.reason ?? null;
}

// GET /accounts
router.get("/", (req, res) => {
  const accounts = readConfig();
  res.json(accounts);
});

// POST /accounts
router.post("/", (req, res) => {
  const { friendlyName, discordId, steam64ids, customItems = [] } = req.body;

  if (!friendlyName || !discordId || !Array.isArray(steam64ids) || steam64ids.length === 0) {
    return res.status(400).json({ error: "friendlyName, discordId, and steam64ids[] are required" });
  }
  if (steam64ids.length > MAX_STEAM64IDS) {
    return res.status(400).json({ error: `Too many steam64ids (max ${MAX_STEAM64IDS})` });
  }
  const invalidId = steam64ids.find(id => !isValidSteam64id(id));
  if (invalidId) {
    return res.status(400).json({ error: `Invalid steam64id: ${invalidId}` });
  }
  if (customItems.length > MAX_CUSTOM_ITEMS) {
    return res.status(400).json({ error: `Too many customItems (max ${MAX_CUSTOM_ITEMS})` });
  }
  for (const id of steam64ids) {
    const reason = getBadEntryReason('steam64id', id);
    if (reason) return res.status(400).json({ error: `steam64id ${id} was previously rejected: ${reason}` });
  }
  for (const item of customItems) {
    const reason = getBadEntryReason('item', item);
    if (reason) return res.status(400).json({ error: `item "${item}" was previously rejected: ${reason}` });
  }

  const accounts = readConfig();
  if (accounts.find((a) => a.discordId === discordId)) {
    return res.status(409).json({ error: "Account with this discordId already exists" });
  }

  const uid = crypto.randomBytes(8).toString("hex");
  const account = { uid, friendlyName, discordId, steam64ids, customItems };
  accounts.push(account);
  writeConfig(accounts);

  for (const id of steam64ids) enqueueInventoryIfDue(id);
  for (const item of customItems) enqueuePrice(item);

  logger.info({ uid, friendlyName, discordId }, "Account added");
  res.status(201).json(account);
});

// POST /accounts/discord — create account via Discord, return uid
// If discordId already exists, returns 409 with the existing uid
router.post("/discord", (req, res) => {
  const { discordId, friendlyName } = req.body;

  if (!discordId) {
    return res.status(400).json({ error: "discordId is required" });
  }

  const accounts = readConfig();
  const existing = accounts.find((a) => a.discordId === discordId);
  if (existing) {
    return res.status(409).json({ error: "Account with this discordId already exists" });
  }

  const uid = crypto.randomBytes(8).toString("hex");
  const account = { uid, friendlyName: friendlyName || null, discordId, steam64ids: [], customItems: [] };
  accounts.push(account);
  writeConfig(accounts);

  logger.info({ uid, discordId }, "Account created via Discord");
  res.status(201).json({ uid });
});

// GET /accounts/:uid
router.get("/:uid", (req, res) => {
  const { account } = getAccount(req.params.uid);
  if (!account) return res.status(404).json({ error: "Account not found" });
  res.json(account);
});

// PUT /accounts/:uid
router.put("/:uid", (req, res) => {
  const { accounts, account } = getAccount(req.params.uid);
  if (!account) return res.status(404).json({ error: "Account not found" });

  const idx = accounts.findIndex((a) => a.uid === req.params.uid);
  if (req.body.steam64ids !== undefined && req.body.steam64ids.length > MAX_STEAM64IDS) {
    return res.status(400).json({ error: `Too many steam64ids (max ${MAX_STEAM64IDS})` });
  }
  if (req.body.steam64ids !== undefined) {
    const invalidId = req.body.steam64ids.find(id => !isValidSteam64id(id));
    if (invalidId) return res.status(400).json({ error: `Invalid steam64id: ${invalidId}` });
  }
  if (req.body.customItems !== undefined && req.body.customItems.length > MAX_CUSTOM_ITEMS) {
    return res.status(400).json({ error: `Too many customItems (max ${MAX_CUSTOM_ITEMS})` });
  }
  if (req.body.steam64ids !== undefined) {
    for (const id of req.body.steam64ids) {
      const reason = getBadEntryReason('steam64id', id);
      if (reason) return res.status(400).json({ error: `steam64id ${id} was previously rejected: ${reason}` });
    }
  }
  if (req.body.customItems !== undefined) {
    for (const item of req.body.customItems) {
      const reason = getBadEntryReason('item', item);
      if (reason) return res.status(400).json({ error: `item "${item}" was previously rejected: ${reason}` });
    }
  }
  if (req.body.friendlyName !== undefined) accounts[idx].friendlyName = req.body.friendlyName;
  if (req.body.discordId !== undefined) accounts[idx].discordId = req.body.discordId;
  if (req.body.steam64ids !== undefined) accounts[idx].steam64ids = req.body.steam64ids;
  if (req.body.customItems !== undefined) accounts[idx].customItems = req.body.customItems;
  writeConfig(accounts);

  if (req.body.steam64ids !== undefined) {
    for (const id of accounts[idx].steam64ids) enqueueInventoryIfDue(id);
  }
  if (req.body.customItems !== undefined) {
    for (const item of accounts[idx].customItems) enqueuePrice(item);
  }

  res.json(accounts[idx]);
});

// DELETE /accounts/:uid
router.delete("/:uid", (req, res) => {
  const { accounts, account } = getAccount(req.params.uid);
  if (!account) return res.status(404).json({ error: "Account not found" });

  const idx = accounts.findIndex((a) => a.uid === req.params.uid);
  accounts.splice(idx, 1);
  writeConfig(accounts);

  logger.info({ uid: account.uid, friendlyName: account.friendlyName }, "Account deleted");
  res.status(204).send();
});

// POST /accounts/:uid/steam64ids — add a steam64id (no-op if already present)
router.post("/:uid/steam64ids", (req, res) => {
  const { accounts, account } = getAccount(req.params.uid);
  if (!account) return res.status(404).json({ error: "Account not found" });
  const { steam64id } = req.body;
  if (!steam64id) return res.status(400).json({ error: "steam64id is required" });
  if (!isValidSteam64id(steam64id)) return res.status(400).json({ error: `Invalid steam64id: ${steam64id}` });
  const badIdReason = getBadEntryReason('steam64id', steam64id);
  if (badIdReason) return res.status(400).json({ error: `steam64id ${steam64id} was previously rejected: ${badIdReason}` });

  const idx = accounts.findIndex(a => a.uid === req.params.uid);
  if (!accounts[idx].steam64ids.includes(steam64id)) {
    if (accounts[idx].steam64ids.length >= MAX_STEAM64IDS) {
      return res.status(400).json({ error: `Too many steam64ids (max ${MAX_STEAM64IDS})` });
    }
    accounts[idx].steam64ids.push(steam64id);
    writeConfig(accounts);
  }
  enqueueInventoryIfDue(steam64id);
  res.json(accounts[idx]);
});

// DELETE /accounts/:uid/steam64ids/:id — remove a steam64id
router.delete("/:uid/steam64ids/:id", (req, res) => {
  const { accounts, account } = getAccount(req.params.uid);
  if (!account) return res.status(404).json({ error: "Account not found" });

  const idx = accounts.findIndex(a => a.uid === req.params.uid);
  const pos = accounts[idx].steam64ids.indexOf(req.params.id);
  if (pos === -1) return res.status(404).json({ error: "steam64id not found on account" });

  accounts[idx].steam64ids.splice(pos, 1);
  writeConfig(accounts);
  res.json(accounts[idx]);
});

// POST /accounts/:uid/customItems — add a custom item (no-op if already present)
router.post("/:uid/customItems", (req, res) => {
  const { accounts, account } = getAccount(req.params.uid);
  if (!account) return res.status(404).json({ error: "Account not found" });
  const { item } = req.body;
  if (!item) return res.status(400).json({ error: "item is required" });
  const badItemReason = getBadEntryReason('item', item);
  if (badItemReason) return res.status(400).json({ error: `item "${item}" was previously rejected: ${badItemReason}` });

  const idx = accounts.findIndex(a => a.uid === req.params.uid);
  if (!accounts[idx].customItems.includes(item)) {
    if (accounts[idx].customItems.length >= MAX_CUSTOM_ITEMS) {
      return res.status(400).json({ error: `Too many customItems (max ${MAX_CUSTOM_ITEMS})` });
    }
    accounts[idx].customItems.push(item);
    writeConfig(accounts);
  }
  enqueuePrice(item);
  res.json(accounts[idx]);
});

// DELETE /accounts/:uid/customItems/:item — remove a custom item
router.delete("/:uid/customItems/:item", (req, res) => {
  const { accounts, account } = getAccount(req.params.uid);
  if (!account) return res.status(404).json({ error: "Account not found" });

  const idx = accounts.findIndex(a => a.uid === req.params.uid);
  const item = decodeURIComponent(req.params.item);
  const pos = accounts[idx].customItems.indexOf(item);
  if (pos === -1) return res.status(404).json({ error: "item not found on account" });

  accounts[idx].customItems.splice(pos, 1);
  writeConfig(accounts);
  res.json(accounts[idx]);
});

// GET /accounts/:uid/inventory — live passthrough to Steam, all steam64ids merged
router.get("/:uid/inventory", async (req, res) => {
  const { account } = getAccount(req.params.uid);
  if (!account) return res.status(404).json({ error: "Account not found" });

  try {
    const results = await Promise.all(account.steam64ids.map((id) => fetchInventory(id)));
    const items = results.flat();
    res.json({ uid: account.uid, count: items.length, items });
  } catch (err) {
    logger.error({ err, uid: account.uid }, "Failed to fetch inventory");
    res.status(502).json({ error: err.message });
  }
});

// GET /accounts/:uid/summary — inventory items per steam64id + custom items, each with latest price
router.get("/:uid/summary", (req, res) => {
  const { account } = getAccount(req.params.uid);
  if (!account) return res.status(404).json({ error: "Account not found" });

  const latestPrice = db.prepare(`
    SELECT ps.lowest_price, ps.median_price, ps.volume, ps.captured_at
    FROM price_snapshots ps
    WHERE ps.item_id = (SELECT id FROM item_names WHERE name = ?)
    ORDER BY ps.captured_at DESC
    LIMIT 1
  `);

  // Inventory items per steam64id
  const invQuery = db.prepare(`
    SELECT n.name AS market_hash_name, ii.first_seen, ii.last_seen, ii.missing
    FROM inventory_items ii
    JOIN item_names n ON n.id = ii.item_id
    WHERE ii.steam64id = ?
    ORDER BY n.name
  `);

  const steam64ids = {};
  for (const id of (account.steam64ids || [])) {
    const rows = invQuery.all(id);
    steam64ids[id] = rows.map(r => ({
      market_hash_name: r.market_hash_name,
      first_seen: r.first_seen,
      last_seen: r.last_seen,
      missing: r.missing === 1,
      price: latestPrice.get(r.market_hash_name) ?? null,
    }));
  }

  // Custom items with latest price
  const customItems = (account.customItems || []).map(name => ({
    market_hash_name: name,
    price: latestPrice.get(name) ?? null,
  }));

  res.json({ uid: account.uid, friendlyName: account.friendlyName, steam64ids, customItems });
});

// GET /accounts/:uid/progress — scan state per steam64id and custom item
router.get("/:uid/progress", (req, res) => {
  const { account } = getAccount(req.params.uid);
  if (!account) return res.status(404).json({ error: "Account not found" });

  const reenqueueDelaySecs = Math.floor(REENQUEUE_DELAY_MS / 1000);

  const lastFetchStmt = db.prepare(`
    SELECT item_count, duration_ms, fetched_at
    FROM inventory_fetches
    WHERE steam64id = ?
    ORDER BY fetched_at DESC
    LIMIT 1
  `);

  const steam64ids = {};
  for (const id of (account.steam64ids || [])) {
    const queued = isInventoryQueued(id);
    const lastFetch = lastFetchStmt.get(id) ?? null;
    steam64ids[id] = {
      queued,
      lastFetch,
      nextScanAt: (!queued && lastFetch) ? lastFetch.fetched_at + reenqueueDelaySecs : null,
    };
  }

  const lastPriceStmt = db.prepare(`
    SELECT ps.lowest_price, ps.captured_at
    FROM price_snapshots ps
    WHERE ps.item_id = (SELECT id FROM item_names WHERE name = ?)
    ORDER BY ps.captured_at DESC
    LIMIT 1
  `);

  const customItems = {};
  for (const name of (account.customItems || [])) {
    const queued = isPriceQueued(name);
    const lastPrice = lastPriceStmt.get(name) ?? null;
    customItems[name] = {
      queued,
      lastPrice,
      nextScanAt: (!queued && lastPrice) ? lastPrice.captured_at + reenqueueDelaySecs : null,
    };
  }

  const { inventoryQueueSize, priceQueueSize } = getQueueState();
  res.json({ uid: account.uid, inventoryQueueSize, priceQueueSize, steam64ids, customItems });
});

// GET /accounts/:uid/prices
router.get("/:uid/prices", (req, res) => {
  const { account } = getAccount(req.params.uid);
  if (!account) return res.status(404).json({ error: "Account not found" });

  const days = parseInt(req.query.days ?? "7", 10);
  const since = Math.floor(Date.now() / 1000) - days * 24 * 60 * 60;
  const itemFilter = req.query.item;

  const items = itemFilter ? [itemFilter] : Array.isArray(account.customItems) ? account.customItems : [];

  if (items.length === 0) return res.json({});

  const placeholders = items.map(() => "?").join(", ");
  const snapshots = db
    .prepare(
      `
    SELECT n.name AS market_hash_name, ps.lowest_price, ps.median_price, ps.volume, ps.captured_at
    FROM price_snapshots ps
    JOIN item_names n ON n.id = ps.item_id
    WHERE ps.item_id IN (SELECT id FROM item_names WHERE name IN (${placeholders})) AND ps.captured_at >= ?
    ORDER BY n.name, ps.captured_at DESC
  `
    )
    .all(...items, since);

  const grouped = {};
  for (const s of snapshots) {
    if (!grouped[s.market_hash_name]) grouped[s.market_hash_name] = [];
    grouped[s.market_hash_name].push({
      lowest_price: s.lowest_price,
      median_price: s.median_price,
      volume: s.volume,
      captured_at: s.captured_at
    });
  }

  res.json(grouped);
});

module.exports = router;